dockerized_dbt/container_python_etl/requirements.txt
psycopg2-binary==2.9.6
pandas==2.2
